feat: flip through output buffers in control panel
Added cursor tracking to the reactive view. This is done by adding an autocommand
to the buffer that checks the cursor position and executes all on_hover callbacks
where it's range contains the current cursor position.

This currently does not check if the cursor just entered the range, so it will
execute the callback every time the cursor moves within the range. This is fine
for now as our only use case is for single lines, but needs to be dealt with
soon~ish.

Added some keybindings to better navigate between output butter and control panel.